move to pkg and fix test paths
