+++ Dockerfile-cni-plugin
@@ -6,7 +6,7 @@ WORKDIR /build
 COPY --link go.mod go.sum ./
 COPY --link ./cni-plugin ./cni-plugin
 COPY --link ./proxy-init ./proxy-init
-COPY --link ./internal ./internal
+COPY --link ./pkg ./pkg
 RUN go mod download
 ARG TARGETARCH
 RUN CGO_ENABLED=0 GOOS=linux GOARCH=$TARGETARCH GO111MODULE=on \
