+++ cni-plugin/integration/Dockerfile-tester
@@ -14,6 +14,6 @@ ENV GOCACHE=/tmp/
 WORKDIR /src
 COPY --link go.mod go.sum .
 COPY --link cni-plugin cni-plugin
-COPY --link internal internal
+COPY --link pkg pkg
 COPY --link proxy-init proxy-init
 RUN go mod tidy && go mod download
